Fix query executor database type mismatch issue
Problem:
- Large queries like rarity:mythic were taking too long
- Second query with same params returned error "Failed to fetch cards by IDs"
- Root cause: cmc column type DECIMAL/NUMERIC incompatible with Rust f64 type

Investigation:
- Added improved error logging to cache manager to distinguish between
  query executor errors and empty results
- Added detailed database error logging to query executor
- Discovered type mismatch: "Rust type Option<f64> is not compatible with SQL type NUMERIC"

Solution:
- Changed cmc column type from DECIMAL to DOUBLE PRECISION in database
- Updated migration 001_initial_schema.sql to use DOUBLE PRECISION
- Created migration 002_fix_cmc_type.sql for existing databases
- Improved error logging in cache manager and query executor

Results:
- Query executor now successfully queries local database for all filters
- Cache system working correctly for large result sets
- Second queries use cache (100 IDs) and return results quickly
- No more fallback to Scryfall API for queries with local data

+++ migrations/001_initial_schema.sql
@@ -9,7 +9,7 @@ CREATE TABLE IF NOT EXISTS cards (
     oracle_id UUID,
     name TEXT NOT NULL,
     mana_cost TEXT,
-    cmc DECIMAL,
+    cmc DOUBLE PRECISION,
     type_line TEXT,
     oracle_text TEXT,
     colors TEXT[],
